LOC-2 Removed room password from submission and from input format of /room create
New syntax is

	/room create [roomname] [owner]

The client will then generate _client side_ a password to use as the room password. It's perfectly fine to use a different value, so long as all other users in the room are using the same one.

What _will_ need to happen when creating a room, though, is for us to submit a password hash for the owning user so that they can actually enter the room to invite others.