LOC-8 Do not throw exception if message cannot be decrypted
If a user connects into a room using a different encryption key to everyone else, do not let them effectively DoS the room by causing exceptions in everyone else's clients.

If we receive a message that cannot be decrypted, we simply print a warning to that effect and carry on.

Because the sending user's name is in the encrypted payload, we can't ascertain who sent it - but as soon as someone else replies, that sending user will also get an error message.