LOC-6 Initial (and very rough) periodic message clearance
At server startup we spin up a seperate scheduler thread, which places a request to the main app every 60 seconds, in order to purge any messages older than the threshold in the system.

It's currently very, very rough, but seems to work.

There's still a bit to be done here though

	- Need to validate the submitted password
	- Need to actually generate a password rather than using the temporarily hardcoded 1234

And probably quite a lot else too