LOC-6 Use autogenerated password for scheduler requests
At startup we generate a 64 char password for the scheduler to identify itself with.

Also, move the scheduler thread launch to within the app processing thread. We do this because otherwise we end up with two scheduler threads (because of the way Flask handles launching the 'main' thread). Also means it'll refresh if the processing thread is changed for some reason.