LOC-11 Add a field to record the last activity in a room
Currently unused by anything, but will be used for the automatic room closure.

Only gets updated when a message is sent by a user. May want to think about whether or not we want to also update when system messages are sent - it probably makes sense to, as it'd be irritating if you joined seconds before the closure threshold and hadn't sent a message yet.