LOC-15 Implement Testing mode
When the argument '--enable-testing-mode' is included in the server's commandline, the database will be created on disk instead of in memory.

This is so that tests can check the database directly to verify that what's been written into the database is what's expected.

When testing mode is enabled, the server will print a warning on the commandline at startup.
It will also push a message into rooms whenever a new user joins, so that if it is inadvertantly used, users will know that their message payloads are being written to disk.