LOC-24 Build some basic tests
As much to start getting the structure as anything - it could be better/tidier but it's functional.

Uses the client as a module in order to run 3 basic tests

- test_one Create a room
- test_two Attempt to join with invalid credentials
- test_three Join with valid credentials

Test functions simply return a list of the following structure

	{noformat}
	[
	{'Test' : 'Join the room','Result' : 'FAIL', 'Notes': '' },
	isFatal
	]
	{noformat}

Where {{isFatal}} dictates whether a failure should abort the test run or not (i.e. if other tests depend on something it does, just give up)